+++ ijkmedia/ijkplayer/ff_ffplay_options.h
@@ -151,8 +151,12 @@ static const AVOption ffp_context_options[] = {
 
     { "enable-accurate-seek",                      "enable accurate seek",
         OPTION_OFFSET(enable_accurate_seek),       OPTION_INT(0, 0, 1) },
+    { "accurate-seek-timeout",                      "accurate seek timeout",
+        OPTION_OFFSET(accurate_seek_timeout),       OPTION_INT(MAX_ACCURATE_SEEK_TIMEOUT, 0, MAX_ACCURATE_SEEK_TIMEOUT) },
     { "skip-calc-frame-rate",                      "don't calculate real frame rate",
         OPTION_OFFSET(skip_calc_frame_rate),       OPTION_INT(0, 0, 1) },
+    { "get-frame-mode",                      "warning, this option only for get frame",
+        OPTION_OFFSET(get_frame_mode),       OPTION_INT(0, 0, 1) },
 
         // iOS only options
     { "videotoolbox",                       "VideoToolbox: enable",